[Fixed] 修改app_nofification中的时间字段类型为bigint

+++ webfast-kernel-dao-impl-jdbc/src/main/resources/sql/mysql/webfast-kernel.sql
@@ -187,11 +187,11 @@ CREATE TABLE `app_navigation` (
 -- ----------------------------
 DROP TABLE IF EXISTS `app_notification`;
 CREATE TABLE `app_notification` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT '通知ID',
-  `userId` int(10) unsigned NOT NULL COMMENT '被通知的用户ID',
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT COMMENT '通知ID',
+  `userId` int(11) unsigned NOT NULL COMMENT '被通知的用户ID',
   `type` varchar(64) NOT NULL DEFAULT 'default' COMMENT '通知类型',
   `content` text COMMENT '通知内容',
-  `createdTime` int(10) unsigned NOT NULL COMMENT '通知时间',
+  `createdTime` bigint unsigned NOT NULL COMMENT '通知时间',
   `isRead` tinyint(1) NOT NULL DEFAULT '0' COMMENT '是否已读',
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT='用户通知表';
